+++ docs/release_notes.rst
@@ -15,6 +15,8 @@ New Features
 
 * Extend support for detecting missing arguments in Google style
   docstrings to method calls (#384).
+* Extend support for detecting missing argument description in Numpy style
+  docstrings (#407).
 * Added support for Python 3.8 (#423).
 
 Bug Fixes
@@ -25,6 +27,7 @@ Bug Fixes
   backslashes for line continuation or unicode literals ``\u...`` and
   ``\N...`` anymore. These are considered intended elements of the docstring
   and thus should not be escaped by using a raw docstring (#365).
+* Fix decorator parsing (#411).
 
 4.0.1 - August 14th, 2019
 -------------------------
@@ -39,7 +42,7 @@ Bug Fixes
   the violation code table (#396).
 * Fixed IndentationError when parsing function arguments (#392).
 
-  
+
 4.0.0 - July 6th, 2019
 ----------------------